+++ test/Client/Main.js
@@ -2,6 +2,5 @@ import "./DateExtensionsTest.js";
 import "./FileExtensionsTest.js";
 import "./NumberExtensionsTest.js";
 import "./StringExtensionsTest.js";
-// import "./Data/PaginationTest.js";
+import "./Data/PaginationTest.js";
 // import "./Data/SortTest.js";
-// import "./DependencyInjection/ContainerTest.js";